Déplace les formulaires relatifs aux auteurs avec leurs vues
