+++ zds/tutorialv2/forms.py
@@ -13,7 +13,6 @@ from zds.utils.models import SubCategory
 from zds.tutorialv2.models import TYPE_CHOICES
 from zds.tutorialv2.models.database import PublishableContent
 from django.utils.translation import gettext_lazy as _
-from zds.member.models import Profile
 from zds.utils.forms import IncludeEasyMDE
 from zds.utils.validators import with_svg_validator, slugify_raise_on_invalid, InvalidSlugError
 
@@ -45,63 +44,6 @@ class FormWithTitle(forms.Form):
 class ReviewerTypeModelChoiceField(forms.ModelChoiceField):
     def label_from_instance(self, obj):
         return obj.title
-
-
-class AuthorForm(forms.Form):
-    username = forms.CharField(label=_("Auteurs à ajouter séparés d'une virgule."), required=True)
-
-    def __init__(self, *args, **kwargs):
-        super().__init__(*args, **kwargs)
-        self.helper = FormHelper()
-        self.helper.form_class = "content-wrapper"
-        self.helper.form_method = "post"
-        self.helper.layout = Layout(
-            Field("username"),
-            ButtonHolder(
-                StrictButton(_("Ajouter"), type="submit"),
-            ),
-        )
-
-    def clean_username(self):
-        """Check every username and send it to the cleaned_data['user'] list
-
-        :return: a dictionary of all treated data with the users key added
-        """
-        cleaned_data = super().clean()
-        users = []
-        if cleaned_data.get("username"):
-            for username in cleaned_data.get("username").split(","):
-                user = (
-                    Profile.objects.contactable_members()
-                    .filter(user__username__iexact=username.strip().lower())
-                    .first()
-                )
-                if user is not None:
-                    users.append(user.user)
-            if len(users) > 0:
-                cleaned_data["users"] = users
-        return cleaned_data
-
-    def is_valid(self):
-        return super().is_valid() and "users" in self.clean()
-
-
-class RemoveAuthorForm(AuthorForm):
-    def clean_username(self):
-        """Check every username and send it to the cleaned_data['user'] list
-
-        :return: a dictionary of all treated data with the users key added
-        """
-        cleaned_data = super(AuthorForm, self).clean()
-        users = []
-        for username in cleaned_data.get("username").split(","):
-            # we can remove all users (bots inclued)
-            user = Profile.objects.filter(user__username__iexact=username.strip().lower()).first()
-            if user is not None:
-                users.append(user.user)
-        if len(users) > 0:
-            cleaned_data["users"] = users
-        return cleaned_data
 
 
 class ContainerForm(FormWithTitle):
